Rename test setup module to testUtil
